Move hook to kubeUtils
