+++ packages/manager/src/features/Kubernetes/CreateCluster/CreateCluster.tsx
@@ -19,8 +19,8 @@ import {
   getLatestVersion,
   useAPLAvailability,
   useIsLkeEnterpriseEnabled,
+  useLkeStandardOrEnterpriseVersions,
 } from 'src/features/Kubernetes/kubeUtils';
-import { useLkeStandardOrEnterpriseVersions } from 'src/hooks/useLkeStandardOrEnterpriseVersions';
 import { useAccount } from 'src/queries/account/account';
 import {
   reportAgreementSigningError,
